merge master

+++ storage/cleanup.py
@@ -4,7 +4,7 @@ from db_ingest import DBIngest
 from db_query import DBQuery
 from signal import signal, SIGTERM
 from language import text
-import time
+from threading import Event
 import os
 
 retention_time=float(os.environ["RETENTION_TIME"])  # in seconds
@@ -14,35 +14,31 @@ office=list(map(float, os.environ["OFFICE"].split(","))) if "OFFICE" in os.envir
 dbhost=os.environ["DBHOST"]
 storage="/var/www/mp4"
 
-dbs=None
-rs=None
-
+stop=Event()
 def quit_service(signum, sigframe):
-    if dbs and rs: dbs.delete(rs["_id"])
-    exit(143)
+    stop.set()
 
 signal(SIGTERM, quit_service)
 dbs=DBIngest(index="services",office=office,host=dbhost)
-while isinstance(office,list):
-    try:
-        rs=dbs.ingest({
-            "name": text["cleanup"],
-            "service": text["maintenance"],
-            "status": "active",
-        })
-        break
-    except Exception as e:
-        print("Waiting for DB...", flush=True)
-        time.sleep(10)
+rs=None
+if isinstance(office,list):
+    dbs.wait(stop)
+    rs=dbs.ingest({
+        "name": text["cleanup"],
+        "service": text["maintenance"],
+        "status": "active",
+    })
 
-while True: 
-    time.sleep(service_interval)
-
+while not stop.is_set(): 
     print("Searching...",flush=True)
     for index in indexes:
+        if stop.is_set(): break
+
         db=DBQuery(index=index,office=office,host=dbhost)
         try:
             for r in db.search("time<now-"+str(retention_time*1000), size=500):
+                if stop.is_set(): break
+
                 # delete the record
                 db.delete(r["_id"])
     
@@ -55,5 +51,8 @@ while True:
                         pass
         except Exception as e:
             print("Exception: "+str(e), flush=True)
-            time.sleep(10)
 
+    stop.wait(service_interval)
+
+if rs:
+    dbs.delete(rs["_id"])
